+++ media/js/main-scripts.js
@@ -81,6 +81,12 @@ $(function() {
     });
   }
   
+  CKEDITOR.on('instanceCreated', function(ev) {
+    ev.editor.on('resize',function(reEvent) {
+      if($item = $('#item')) { itemMasonry(); }
+    });
+  });
+  
   $(this).on('mouseup', "#copy", function() {
     $(this).select();
   });
@@ -99,7 +105,7 @@ $(function() {
   if($('textarea[name="article"]').length)
   {
     $item = $("#item");
-    //$('textarea[name="article"]').wysibb(wbbOpt);
+    $('textarea[name="article"]').ckeditor();
     var timer = setInterval(function() {
       if($("#cke_item-article").height()) {
         itemMasonry();
@@ -440,7 +446,7 @@ $(function() {
       modal: true,
       open: function( event, ui ) {
         if(!navigator.userAgent.match(browsers)) {
-          //$('textarea[name="article"]').wysibb(wbbOpt);
+          $('textarea[name="article"]').ckeditor();
           var timer = setInterval(function() {
             if($("#cke_item-article").height()) {
               itemMasonry();
@@ -563,7 +569,7 @@ $(function() {
       modal: true,
       open: function( event, ui ) {
         if(!navigator.userAgent.match(browsers)) {
-          //$('textarea[name="article"]').wysibb(wbbOpt);
+          $('textarea[name="article"]').ckeditor();
           var timer = setInterval(function() {
             if($("#cke_item-article").height()) {
               itemMasonry();
@@ -961,7 +967,7 @@ $(function() {
               }
               if($('textarea[name="article"]').length)
               {
-                //$('textarea[name="article"]').wysibb(wbbOpt);
+                $('textarea[name="article"]').ckeditor();
                 var timer2 = setInterval(function() {
                   if($("#cke_item-article").height()) {
                     itemMasonry();